Pull out IsLedger from Ledger.Abstract
This allows us to define the `TopLevelConfig` independently from `ApplyBlock`,
which is then required to _pass_ the `TopLevelConfig` _to_ `ApplyBlock`.
