+++ ouroboros-consensus/ouroboros-consensus.cabal
@@ -101,6 +101,7 @@ library
                        Ouroboros.Consensus.HardFork.History.Util
                        Ouroboros.Consensus.HeaderValidation
                        Ouroboros.Consensus.Ledger.Abstract
+                       Ouroboros.Consensus.Ledger.Basics
                        Ouroboros.Consensus.Ledger.CommonProtocolParams
                        Ouroboros.Consensus.Ledger.Dual
                        Ouroboros.Consensus.Ledger.Extended
